refactor: renamed and moved tasks and create Deobfuscate task to split the creation
Took 6 minutes
